Add files via upload
As the basic law of Physics says, energy can neither be created nor be  destroyed  it  can  only  be  converted  from  one  form  to  another.  During huge  amount  of  energy  is  lost  to  atmosphere  as  heat.  It will  be  good  if  we  could  use  this  energy  somehow  which  is  otherwise  getting  wasted out  and  reuse  it  next time  we  started  to  accelerate.  Regenerative  braking refers  to  a  system  in  which  the  kinetic  energy  of  the  vehicle  is  stored  temporarily,  as  an  accumulative  energy,  during  deceleration,  and  is  reused as  kinetic  energy  during  acceleration  or  running.  Regenerative  braking  is  a  small,  yet  very  important,  step  toward  our  eventual  independence  from fossil  fuels. These  kinds  of  brakes  allow  capacitor  to  be  used  for  temporary periods  of  time  without  the  need  to  be  plugged  into  an  external  charger.  These  types  of  brakes  also  extend  the  driving  range  of  fully  electric vehicles.  Regenerative  braking  is  a  way  to  extend  range  of  the  electric vehicles.  In  many  hybrid   vehicles  cases,  this  system  is  also  applied hybrid  vehicles  to  improve  fuel  economy. A  normal  car  is  only  about 20%  efficient,  meaning  some  80%  of  the  energy  it  expends  is  wasted  as heat  created  by  friction.  
Regenerative  braking  could  reclaim as  much  as  half  of  that  wasted  energy,  which  equates  into  more motion  produced  by  the  fuel  we  are paying  for  instead  of  using  that  fuel to  create  heat  that  is  being  dissipated  uselessly  into  the  environment.